Start generalizing zip proxy code.
